+++ src/Kiota.Builder/GenerationConfiguration.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace Kiota.Builder {
     public class GenerationConfiguration {
@@ -23,10 +24,10 @@ namespace Kiota.Builder {
         };
         private static readonly HashSet<GenerationLanguage> BarreledLanguages = new () {
             GenerationLanguage.Ruby,
-            // TODO: add typescript once we have a barrel writer for it
+            GenerationLanguage.TypeScript
         };
         private static readonly HashSet<GenerationLanguage> BarreledLanguagesWithConstantFileName = new () {
-            //TODO: add typescript once we have a barrel writer for it
+            GenerationLanguage.TypeScript
         };
         public bool CleanOutput { get; set;}
     }