+++ src/Kiota.Builder/KiotaBuilder.cs
@@ -12,6 +12,7 @@ using Kiota.Builder.Extensions;
 using Kiota.Builder.Writers;
 using Microsoft.OpenApi.Any;
 using Kiota.Builder.Refiners;
+using Kiota.Builder.CodeRenderers;
 using System.Security;
 using Microsoft.OpenApi.Services;
 using System.Threading;
@@ -238,7 +239,8 @@ public class KiotaBuilder
         var languageWriter = LanguageWriter.GetLanguageWriter(language, config.OutputPath, config.ClientNamespaceName);
         var stopwatch = new Stopwatch();
         stopwatch.Start();
-        await new CodeRenderer(config).RenderCodeNamespaceToFilePerClassAsync(languageWriter, generatedCode, cancellationToken);
+        var codeRenderer = CodeRenderer.GetCodeRender(config);
+        await codeRenderer.RenderCodeNamespaceToFilePerClassAsync(languageWriter, generatedCode, cancellationToken);
         stopwatch.Stop();
         logger.LogTrace("{timestamp}ms: Files written to {path}", stopwatch.ElapsedMilliseconds, config.OutputPath);
     }