+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- TypeScript adding index exporting models to fix #870.
 - Fixed a bug where JSON serialization would fail on nil properties in Go.
 
 ## [0.0.19] - 2022-03-18